readMsg: Do not always add magic on recFile
The readMsg function always prepends the magic data, even if the actual
message is empty. This particularly goes wrong when in passive mode or
rs485 mode and the first data byte already contains the magic value.

Which in turn is due to the odd way readMsg works in itself, the inFile
pointer is expected to sit just after the magic value.

This is fixed by not recording empty messages (where the header is
re-added again).

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>